requirements.txt
streamlit
ollama
pydantic
httpx
tldextract
python-whois
python-dotenv
markdown
weasyprint
